# openapi.yaml
openapi: 3.1.0
info:
  title: GitHub Issues Gateway
  version: 1.0.0
  summary: A validated HTTP gateway over the GitHub Issues REST API.
  description: |
    A small service that fronts the GitHub Issues REST API for one configured
    repository (`GITHUB_OWNER/GITHUB_REPO`) and ingests that repository's
    webhooks.

    ## Authentication

    Clients of *this* API are unauthenticated -- it is intended to run on
    localhost or behind your own gateway. The service authenticates to GitHub
    on the caller's behalf using a fine-grained PAT supplied out-of-band in
    `GITHUB_TOKEN`; that is the `githubToken` security scheme below. It is
    declared on the routes that reach GitHub so the dependency is explicit in
    the contract, but it is **never** accepted from the client.

    `POST /webhook` is authenticated differently: GitHub signs each delivery
    with an HMAC-SHA256 of the raw body under `WEBHOOK_SECRET`, carried in
    `X-Hub-Signature-256`. That is the `webhookSignature` scheme.

    ## Deleting issues

    GitHub's REST API has no "delete issue" operation. Deletion is modelled as
    `PATCH /issues/{number}` with `{"state": "closed"}`.

    ## Pagination

    `GET /issues` and `GET /issues/{number}/comments` accept `page` and
    `per_page` (max 100) and return an RFC 8288 `Link` header whose `first`,
    `prev`, `next` and `last` URLs point back at **this** service, preserving
    the caller's filters.

    ## Rate limits

...

  contact:
    name: CMPE 272 HW2
...
servers:
...
paths:
...
  /issues:
    get:
      tags: [issues]
      operationId: listIssues
      summary: List issues
      description: |
        Lists issues for the configured repository, preserving GitHub's
        pagination semantics. Pull requests are filtered out: GitHub serves
        them from the same endpoint, but this is an issues API.

        Supports conditional GET. Send the `ETag` from a previous response back
        as `If-None-Match` to get a `304` and save both bandwidth and GitHub
        rate-limit budget.
      security:
        - githubToken: []
      parameters:
        - $ref: "#/components/parameters/StateFilter"
        - $ref: "#/components/parameters/Labels"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: sort
          in: query
          required: false
          description: Field to sort by.
          schema: { type: string, enum: [created, updated, comments] }
        - name: direction
          in: query
          required: false
          description: Sort direction.
          schema: { type: string, enum: [asc, desc] }
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: A page of issues.
          headers:
            Link: { $ref: "#/components/headers/Link" }
            ETag: { $ref: "#/components/headers/ETag" }
            X-Page: { $ref: "#/components/headers/XPage" }
            X-Per-Page: { $ref: "#/components/headers/XPerPage" }
            X-Cache: { $ref: "#/components/headers/XCache" }
            X-Request-ID: { $ref: "#/components/headers/XRequestId" }
            X-RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
            X-RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Issue" }
              examples:
                page:
                  summary: First page, open issues
                  value:
                    - number: 12
                      id: 2451900001
                      title: Rate limiter drops the Retry-After header
                      body: Reproduced on main at 0f21ac9.
                      state: open
                      state_reason: null
                      labels:
                        - { name: bug, color: d73a4a, description: Something is broken }
                      user: { login: mikechau1, id: 583920, type: User, html_url: "https://github.com/mikechau1", avatar_url: "https://avatars.githubusercontent.com/u/583920?v=4" }
                      assignees: []
                      comments: 2
                      locked: false
                      html_url: "https://github.com/mikechau1/cmpe272-issues-gw/issues/12"
                      created_at: "2026-09-06T18:20:11Z"
                      updated_at: "2026-09-06T19:02:44Z"
                      closed_at: null
        "304":
          description: |
            The client's `If-None-Match` matched the current representation.
            No body is returned.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            X-Cache: { $ref: "#/components/headers/XCache" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/NotConfigured" }
        "504": { $ref: "#/components/responses/UpstreamTimeout" }

    post:
      tags: [issues]
      operationId: createIssue
      summary: Create an issue
      security:
        - githubToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateIssueRequest" }
            examples:
              minimal:
                summary: Title only
                value: { title: Something is broken }
              full:
                summary: With body and labels
                value:
                  title: Rate limiter drops the Retry-After header
                  body: "Reproduced on main at 0f21ac9.\n\nSteps: ..."
                  labels: [bug, gateway]
      responses:
        "201":
          description: The issue was created.
          headers:
            Location:
              description: Path of the newly created issue on **this** service.
              schema: { type: string, examples: ["/issues/12"] }
            X-Request-ID: { $ref: "#/components/headers/XRequestId" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
              examples:
                created:
                  value:
                    number: 12
                    id: 2451900001
                    title: Rate limiter drops the Retry-After header
                    body: Reproduced on main at 0f21ac9.
                    state: open
                    state_reason: null
                    labels:
                      - { name: bug, color: d73a4a, description: Something is broken }
                    user: { login: mikechau1, id: 583920, type: User, html_url: "https://github.com/mikechau1", avatar_url: "https://avatars.githubusercontent.com/u/583920?v=4" }
                    assignees: []
                    comments: 0
                    locked: false
                    html_url: "https://github.com/mikechau1/cmpe272-issues-gw/issues/12"
                    created_at: "2026-09-06T18:20:11Z"
                    updated_at: "2026-09-06T18:20:11Z"
                    closed_at: null
        "400":
          description: The request body failed validation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                missing_title:
                  summary: title is required
                  value:
                    error:
                      code: validation_error
                      message: "Request validation failed -- body.title: Field required"
                      status: 400
                      request_id: 8f14e45fceea167a5a36dedd4bea2543
                      details:
                        - field: body.title
                          message: Field required
                          type: missing
                          input: { body: "no title here" }
                blank_title:
                  summary: whitespace-only title
                  value:
                    error:
                      code: validation_error
                      message: "Request validation failed -- body.title: Value error, title must contain at least one non-whitespace character"
                      status: 400
                      request_id: c9f0f895fb98ab9159f51fd0297e236d
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/NotConfigured" }
        "504": { $ref: "#/components/responses/UpstreamTimeout" }
  /issues/{number}:
    parameters:
      - $ref: "#/components/parameters/IssueNumber"
    get:
      tags: [issues]
      operationId: getIssue
      summary: Get a single issue
      security:
        - githubToken: []
      responses:
        "200":
          description: The issue.
          headers:
            X-Request-ID: { $ref: "#/components/headers/XRequestId" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: |
            No such issue, or `{number}` refers to a pull request. GitHub also
            answers 404 (rather than 403) for resources the token cannot see.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                missing:
                  value:
                    error:
                      code: not_found
                      message: "Not found while fetching issue #999. Either the resource does not exist or the token cannot see it (GitHub returns 404 rather than 403 for private resources). GitHub said: Not Found."
                      status: 404
                      request_id: 45c48cce2e2d7fbdea1afc51c7c6ad26
                      upstream_status: 404
                pull_request:
                  value:
                    error:
                      code: not_an_issue
                      message: "#7 is a pull request, not an issue. This gateway exposes issues only."
                      status: 404
                      request_id: d3d9446802a44259755d38e6d163e820
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/NotConfigured" }
        "504": { $ref: "#/components/responses/UpstreamTimeout" }

    patch:
      tags: [issues]
      operationId: updateIssue
      summary: Update an issue (rename, edit, close, or reopen)
      description: |
        Partial update. Only the keys present in the body are sent upstream, so
        omitting a field leaves it untouched while sending `"body": null`
        explicitly clears it.

        This is also the "delete" operation: `{"state": "closed"}`.
      security:
        - githubToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateIssueRequest" }
            examples:
              rename:
                summary: Rename
                value: { title: "Rate limiter drops Retry-After (confirmed)" }
              close:
                summary: Close -- the "D" in CRUD
                value: { state: closed, state_reason: completed }
              reopen:
                summary: Reopen
                value: { state: open }
              edit_body:
                summary: Edit the body
                value: { body: "Updated repro steps." }
      responses:
        "200":
          description: The updated issue.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
              examples:
                closed:
                  value:
                    number: 12
                    id: 2451900001
                    title: Rate limiter drops the Retry-After header
                    body: Updated repro steps.
                    state: closed
                    state_reason: completed
                    labels: []
                    user: { login: mikechau1, id: 583920, type: User, html_url: "https://github.com/mikechau1", avatar_url: "https://avatars.githubusercontent.com/u/583920?v=4" }
                    assignees: []
                    comments: 2
                    locked: false
                    html_url: "https://github.com/mikechau1/cmpe272-issues-gw/issues/12"
                    created_at: "2026-09-06T18:20:11Z"
                    updated_at: "2026-09-06T20:15:03Z"
                    closed_at: "2026-09-06T20:15:03Z"
        "400":
          description: Invalid body -- unknown state, no fields supplied, or unknown key.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                invalid_state:
                  value:
                    error:
                      code: validation_error
                      message: "Request validation failed -- body.state: Input should be 'open' or 'closed'"
                      status: 400
                      request_id: 6512bd43d9caa6e02c990b0a82652dca
                      details:
                        - field: body.state
                          message: "Input should be 'open' or 'closed'"
                          type: literal_error
                          input: deleted
                empty_patch:
                  value:
                    error:
                      code: validation_error
                      message: "Request validation failed -- body: Value error, provide at least one of: title, body, state, state_reason, labels"
                      status: 400
                      request_id: c20ad4d76fe97759aa27a0c99bff6710
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/NotConfigured" }
        "504": { $ref: "#/components/responses/UpstreamTimeout" }
  /issues/{number}/comments:
    parameters:
      - $ref: "#/components/parameters/IssueNumber"
    get:
      tags: [issues]
      operationId: listComments
      summary: List an issue's comments
      security:
        - githubToken: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: A page of comments.
          headers:
            Link: { $ref: "#/components/headers/Link" }
            X-Page: { $ref: "#/components/headers/XPage" }
            X-Per-Page: { $ref: "#/components/headers/XPerPage" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/NotConfigured" }
        "504": { $ref: "#/components/responses/UpstreamTimeout" }

    post:
      tags: [issues]
      operationId: createComment
      summary: Comment on an issue
      security:
        - githubToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommentRequest" }
            examples:
              simple:
                value: { body: "Confirmed on staging -- see the attached trace." }
      responses:
        "201":
          description: The comment was created.
          headers:
            Location:
              description: The comment collection for this issue on **this** service.
              schema: { type: string, examples: ["/issues/12/comments"] }
            X-Request-ID: { $ref: "#/components/headers/XRequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
              examples:
                created:
                  value:
                    id: 3310028841
                    body: "Confirmed on staging -- see the attached trace."
                    user: { login: mikechau1, id: 583920, type: User, html_url: "https://github.com/mikechau1", avatar_url: "https://avatars.githubusercontent.com/u/583920?v=4" }
                    html_url: "https://github.com/mikechau1/cmpe272-issues-gw/issues/12#issuecomment-3310028841"
                    issue_url: "https://api.github.com/repos/mikechau1/cmpe272-issues-gw/issues/12"
                    created_at: "2026-09-06T19:02:44Z"
                    updated_at: "2026-09-06T19:02:44Z"
        "400":
          description: Invalid body.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                empty_body:
                  value:
                    error:
                      code: validation_error
                      message: "Request validation failed -- body.body: String should have at least 1 character"
                      status: 400
                      request_id: aab3238922bcc25a6f606eb525ffdc56
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/NotConfigured" }
        "504": { $ref: "#/components/responses/UpstreamTimeout" }
...
components:
  securitySchemes:
    githubToken:
      type: http
      scheme: bearer
      bearerFormat: GitHub fine-grained PAT
      description: |
        The server-side credential this gateway uses to call GitHub, supplied
        via the `GITHUB_TOKEN` environment variable. Declared so the dependency
        is visible in the contract; it is **never** read from the client
        request.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: |
        HMAC-SHA256 of the raw request body keyed with `WEBHOOK_SECRET`,
        formatted as `sha256=<hex>` and compared in constant time.
  parameters:
    IssueNumber:
      name: number
      in: path
      required: true
      description: Issue number within the repository (not the global issue id).
      schema: { type: integer, minimum: 1, examples: [12] }
    StateFilter:
      name: state
      in: query
      required: false
      description: Filter by issue state.
      schema: { type: string, enum: [open, closed, all], default: open }
    Labels:
      name: labels
      in: query
      required: false
      description: Comma-separated label names; a matching issue carries all of them.
      schema: { type: string, examples: ["bug,gateway"] }
    Page:
      name: page
      in: query
      required: false
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      required: false
      description: Results per page. GitHub's ceiling of 100 is enforced here too.
      schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: A previously returned `ETag`; matching yields `304 Not Modified`.
      schema: { type: string, examples: ['W/"gw1-a1b2c3d4e5f6"'] }
  headers:
    Link:
      description: RFC 8288 pagination links rewritten to point at this service.
      schema:
        type: string
        examples: ['</issues?state=open&page=2&per_page=30>; rel="next", </issues?state=open&page=5&per_page=30>; rel="last"']
    ETag:
      description: Entity tag for conditional GET; send it back as `If-None-Match`.
      schema: { type: string, examples: ['W/"gw1-a1b2c3d4e5f6"'] }
    XPage:
      description: The page number that was served.
      schema: { type: integer, examples: [1] }
    XPerPage:
      description: The page size that was applied after clamping.
      schema: { type: integer, examples: [30] }
    XCache:
      description: Whether the upstream body came from the conditional-GET cache.
      schema: { type: string, enum: [HIT, MISS] }
    XRequestId:
      description: Correlation id, echoed from the request or generated.
      schema: { type: string, examples: [8f14e45fceea167a5a36dedd4bea2543] }
    RateLimitLimit:
      description: GitHub's rate-limit ceiling for the credential in use.
      schema: { type: integer, examples: [5000] }
    RateLimitRemaining:
      description: Requests remaining in the current GitHub rate-limit window.
      schema: { type: integer, examples: [4993] }
    RateLimitReset:
      description: Unix epoch second at which GitHub's window resets.
      schema: { type: integer, examples: [1757203200] }
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: { type: integer, examples: [60] }
  schemas:
...
    Issue:
      type: object
      description: A GitHub issue, projected onto this API's stable surface.
      required: [number, title, state, labels, html_url, created_at, updated_at]
      properties:
        number: { type: integer, description: Issue number within the repository., examples: [12] }
        id: { type: [integer, "null"], description: GitHub's global issue id. }
        title: { type: string }
        body: { type: [string, "null"] }
        state: { type: string, enum: [open, closed] }
        state_reason:
          type: [string, "null"]
          enum: [completed, not_planned, reopened, duplicate, null]
        labels:
          type: array
          items: { $ref: "#/components/schemas/Label" }
        user: { oneOf: [{ $ref: "#/components/schemas/User" }, { type: "null" }] }
        assignees:
          type: array
          description: Assignee logins.
          items: { type: string }
        comments: { type: [integer, "null"], description: Comment count. }
        locked: { type: [boolean, "null"] }
        html_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        closed_at: { type: [string, "null"], format: date-time }
...
  responses:
    BadRequest:
      description: The request failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            invalid_state:
              value:
                error:
                  code: validation_error
                  message: "Request validation failed -- query.state: Input should be 'open', 'closed' or 'all'"
                  status: 400
                  request_id: 1679091c5a880faf6fb5e6087eb1b2dc
                  details:
                    - field: query.state
                      message: "Input should be 'open', 'closed' or 'all'"
                      type: literal_error
                      input: sideways
            per_page_too_large:
              value:
                error:
                  code: validation_error
                  message: "Request validation failed -- query.per_page: Input should be less than or equal to 100"
                  status: 400
                  request_id: 8f14e45fceea167a5a36dedd4bea2543
    Unauthorized:
      description: |
        The gateway has no usable GitHub credential, or GitHub rejected it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            missing_token:
              value:
                error:
                  code: missing_credentials
                  message: "GITHUB_TOKEN is not set, so the gateway has no credentials to call GitHub with. Set a fine-grained PAT with 'Issues: Read and write' on the configured repository."
                  status: 401
                  request_id: e4da3b7fbbce2345d7772b0674a318d5
            rejected_token:
              value:
                error:
                  code: github_unauthorized
                  message: "GitHub rejected the configured GITHUB_TOKEN. Check that the token is valid, unexpired, and scoped to this repository. GitHub said: Bad credentials."
                  status: 401
                  request_id: 1c383cd30b7c298ab50293adfecb7b18
                  upstream_status: 401
    Forbidden:
      description: |
        The token is valid but lacks `Issues: Read and write`, or the
        repository has issues disabled.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            insufficient_permission:
              value:
                error:
                  code: github_forbidden
                  message: "GitHub denied the request. The token is valid but lacks the required permission (Issues: Read and write) on this repository, or the repository has issues disabled. GitHub said: Resource not accessible by personal access token."
                  status: 403
                  request_id: aab3238922bcc25a6f606eb525ffdc56
                  upstream_status: 403
...
    RateLimited:
      description: |
        GitHub's rate limit is exhausted. `Retry-After` says how long to wait.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
        X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
        X-RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            primary:
              value:
                error:
                  code: rate_limited
                  message: "GitHub primary rate limit exhausted while listing issues (limit 5000/hour). Retry after 1800s. GitHub said: API rate limit exceeded."
                  status: 429
                  request_id: d3d9446802a44259755d38e6d163e820
                  upstream_status: 403
    UpstreamError:
      description: GitHub returned 5xx, or was unreachable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            github_down:
              value:
                error:
                  code: github_unavailable
                  message: "GitHub returned 503 while listing issues and the request could not be completed. This is an upstream fault; retry shortly."
                  status: 502
                  request_id: 6512bd43d9caa6e02c990b0a82652dca
                  upstream_status: 503
...